clean the repo for the first release + create tests
