+++ mobility/__init__.py
@@ -1,3 +1,1 @@
-from .mobility_model import MobilityModel, WorkMobilityModel, ShopsMobilityModel
 from .trip_sampler import TripSampler
-from .utilities import read_parquet
