+++ mobility/get_survey_data.py
@@ -4,7 +4,7 @@ import pandas as pd
 import os
 from pathlib import Path
 
-from parsers import prepare_entd_2008, prepare_emp_2019
+from mobility.parsers import prepare_entd_2008, prepare_emp_2019
 
 def get_survey_data(source="EMP-2019"):
     """
