--- mobility/test_get_survey_data.py
@@ -1,3 +1,0 @@
-from get_survey_data import get_survey_data
-
-df = get_survey_data("ENTD-2008")
