+++ test/test_parsers.py
@@ -1,0 +1,7 @@
+from mobility.get_survey_data import get_survey_data
+
+def test_parser_emp_2019():
+	df = get_survey_data("EMP-2019")
+
+def test_parser_entd_2008():
+	df = get_survey_data("ENTD-2008")
